chore: electron folder structure
