+++ package.json
@@ -2,9 +2,9 @@
   "name": "meetalva",
   "version": "0.7.0",
   "description": "Alva is a radically new design tool that enables cross-functional teams to design digital products.",
-  "main": "./build/electron/electron.js",
+  "main": "./build/electron/main.js",
   "scripts": {
-    "build": "tslint --project . -c tslint.json 'src/**/*.ts' && tsc --project . && cp src/electron/*.html build/electron/",
+    "build": "tslint --project . -c tslint.json 'src/**/*.ts' && tsc --project . && cp src/electron/renderer/*.html build/electron/renderer",
     "build:clean": "npm run clean && npm run build",
     "build:electron": "electron-builder -mwl && node ./scripts/dist-postprocess.js",
     "build:lsg": "tsc --project src/lsg --outDir build/lsg/patterns",
@@ -14,7 +14,7 @@
     "dist": "node ./scripts/dist-preprocess.js",
     "docs": "typedoc src",
     "lint": "tslint --project .",
-    "start": "npm run build && concurrently \"electron build/electron/electron.js --remote-debugging-port=9222 --inspect=9223\" \"tsc --project . --watch\"",
+    "start": "npm run build && concurrently \"electron build/electron/main.js --remote-debugging-port=9222 --inspect=9223\" \"tsc --project . --watch\"",
     "start:lsg": "npm run build:lsg && concurrently \"npm run build:lsg -- -w\" \"patternplate start\"",
     "precommit": "lint-staged"
   },
@@ -108,6 +108,7 @@
     "electron-builder": "20.0.5",
     "electron-devtools-installer": "^2.2.1",
     "husky": "^0.14.3",
+    "isomorphic-fetch": "^2.2.1",
     "lint-staged": "^7.0.0",
     "mobx-react-devtools": "^4.2.15",
     "prettier": "^1.9.2",
@@ -122,7 +123,6 @@
     "electron-log": "2.2.14",
     "electron-updater": "2.21.0",
     "fs-extra": "^5.0.0",
-    "isomorphic-fetch": "^2.2.1",
     "js-yaml": "3.11.0",
     "lodash": "^4.17.5",
     "mime-types": "^2.1.18",
