--- src/component/presentation/react/render.tsx
@@ -1,11 +1,0 @@
-import { PreviewApp } from './preview';
-import { HighlightElementFunction } from '../../preview';
-import * as React from 'react';
-import * as ReactDom from 'react-dom';
-
-export const renderReact = (highlightElement: HighlightElementFunction) => {
-	ReactDom.render(
-		<PreviewApp highlightElement={highlightElement} />,
-		document.getElementById('app')
-	);
-};
