+++ src/electron/renderer/app.html
@@ -9,7 +9,7 @@
 		})
 	</script>
 	<div id="app"></div>
-	<script>require('../../build/component/app.js')</script>
+	<script>require('../../../build/electron/renderer/app.js')</script>
 </body>
 
 </html>
