+++ src/styleguide/analyzer/directory.ts
