+++ src/styleguide/analyzer/typescript/export.ts
