+++ src/styleguide/analyzer/typescript/type.ts
