+++ src/styleguide/analyzer/typescript/typescript-internals.d.ts
